b2b-platform/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-filter==23.2
django-cors-headers==4.1.0
psycopg2-binary==2.9.10
Pillow==11.3.0
pandas==2.2.2
openpyxl==3.1.2
drf-spectacular==0.27.2
python-decouple==3.8
dj-database-url==2.1.0
pytest==8.2.1
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.5.1
black==24.4.2
isort==5.13.2
flake8==7.0.0
